requirements.txt
PyYAML==6.0
Jinja2==3.1.2
kubernetes==25.3.0
docopt==0.6.2